feat: add endpoints for all tables

+++ FinalProject/api/models/orders.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String, DECIMAL, DATETIME, Double
+from sqlalchemy import Column, ForeignKey, Integer, String, DECIMAL, DATETIME, Double, Float
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from ..dependencies.database import Base
@@ -10,7 +10,7 @@ class Order(Base):
     tracking_num = Column(Integer, primary_key=True, index=True, autoincrement=True)
     status = Column(String(50))
     date = Column(DATETIME, nullable=False, server_default=str(datetime.now()))
-    price = Column(DECIMAL)
+    price = Column(Float)
     details_link = Column(String(200))
     customer_id = Column(Integer, ForeignKey("customers.id"))
